修改iobug和硬盘同步IO

+++ src/include/phinix/io.h
@@ -4,7 +4,7 @@
 #include <phinix/types.h>
 
 extern u8 in_byte(u16 port); // 输入一个字节
-extern u8 in_word(u16 port); // 输入一个字
+extern u16 in_word(u16 port); // 输入一个字
 
 extern u8 out_byte(u16 port, u8 value); // 输出一个字节
 extern u8 out_word(u16 port, u16 value); // 输出一个字
